chore(trading): move common types to suite-common
