+++ packages/suite/src/views/wallet/trading/common/TradingForm/TradingFormOfferItem.tsx
@@ -1,9 +1,8 @@
-import type { TradingTradeType } from '@suite-common/trading';
+import type { TradingTradeType, TradingUtilsProvidersProps } from '@suite-common/trading';
 import { Row, Spinner, Card, Paragraph } from '@trezor/components';
 import { spacings } from '@trezor/theme';
 
 import { Translation } from 'src/components/suite';
-import { TradingUtilsProvidersProps } from 'src/types/trading/trading';
 import { TradingUtilsProvider } from 'src/views/wallet/trading/common/TradingUtils/TradingUtilsProvider';
 
 interface TradingFormOfferItemProps {
