+++ packages/suite/src/views/wallet/trading/common/TradingUtils/TradingUtilsProvider.tsx
@@ -2,10 +2,9 @@ import styled from 'styled-components';
 
 import { spacings } from '@trezor/theme';
 import { Row } from '@trezor/components';
-import { invityAPI } from '@suite-common/trading';
+import { invityAPI, type TradingUtilsProvidersProps } from '@suite-common/trading';
 
 import { Translation } from 'src/components/suite';
-import { TradingUtilsProvidersProps } from 'src/types/trading/trading';
 
 const Icon = styled.img`
     flex: none;
